+++ src/modules/v1/incoming-gate-pass/incoming-gate-pass.model.ts
@@ -13,6 +13,7 @@ export enum GatePassType {
   RECEIPT = "RECEIPT",
   DELIVERY = "DELIVERY",
   RESTORE = "RESTORE",
+  TRANSFER = "TRANSFER",
 }
 
 /* =======================